+++ testproject/Assets/Tests/Runtime/Animation/AnimatorTestHelper.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 using Unity.Netcode;
@@ -34,6 +35,9 @@ namespace TestProject.RuntimeTests
             {
                 Debug.Log($"[AnimatorTestHelper][{IsServer}] {NetworkManager.name}");
             }
+
+            m_Animator = GetComponent<Animator>();
+            m_NetworkAnimator = GetComponent<NetworkAnimator>();
             if (IsServer)
             {
                 ServerSideInstance = this;
@@ -49,8 +53,14 @@ namespace TestProject.RuntimeTests
             base.OnNetworkSpawn();
         }
 
+        public Action<bool, bool> OnCheckIsServerIsClient;
+
         public override void OnNetworkDespawn()
         {
+            // This verifies the issue where IsServer and IsClient were
+            // being reset prior to NetworkObjects being despawned during
+            // the shutdown period.
+            OnCheckIsServerIsClient?.Invoke(IsServer, IsClient);
             if (ClientSideInstances.ContainsKey(NetworkManager.LocalClientId))
             {
                 ClientSideInstances.Remove(NetworkManager.LocalClientId);
@@ -101,6 +111,16 @@ namespace TestProject.RuntimeTests
         public void SetTrigger(string name = "TestTrigger")
         {
             m_NetworkAnimator.SetTrigger(name);
+        }
+
+        public void SetLateJoinParam(bool isEnabled)
+        {
+            m_Animator.SetBool("LateJoinTest", isEnabled);
+        }
+
+        public Animator GetAnimator()
+        {
+            return m_Animator;
         }
     }
 }