+++ testproject/Assets/Tests/Runtime/MultiprocessRuntime/ExecuteStepInContext.cs
@@ -11,6 +11,7 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using Unity.Netcode.MultiprocessRuntimeTests;
 using Debug = UnityEngine.Debug;
+using Unity.Netcode.MultiprocessRuntimeTests;
 
 /// <summary>
 /// Allows for context based delegate execution.
@@ -84,6 +85,7 @@ public class ExecuteStepInContext : CustomYieldInstruction
 
     internal static void InitializeAllSteps()
     {
+        MultiprocessLogger.Log("InitializeAllSteps - Start");
         // registering magically all context based steps
         IsRegistering = true;
         var registeredMethods = typeof(TestCoordinator).Assembly.GetTypes().SelectMany(t => t.GetMethods())
@@ -148,6 +150,7 @@ public class ExecuteStepInContext : CustomYieldInstruction
 
         IsRegistering = false;
         HasRegistered = true;
+        MultiprocessLogger.Log("InitializeAllSteps - Done");
     }
 
     /// <summary>
@@ -198,10 +201,10 @@ public class ExecuteStepInContext : CustomYieldInstruction
             // MultiprocessLogger.Log($"Registering: {currentActionId}");
             Assert.That(AllActions, Does.Not.Contain(currentActionId)); // sanity check
             AllActions[currentActionId] = this;
+            MultiprocessLogger.Log($"InitializeAllSteps - Registering {currentActionId}");
         }
         else
         {
-            MultiprocessLogger.Log($"Not Registering, m_IsRegistering {m_IsRegistering}: {currentActionId}");
             if (shouldExecuteLocally)
             {
                 m_StepToExecute.Invoke(paramToPass);